Set up the react stuff and moved profile page into own file
Also, added a dummy login page that just writes some text to the screen - to be expanded upon later.

Just a note: a lot of the styling of components that was done within the return() in ProfilePage.js can be done in "react style" - I shall show what I mean soon.