Added display of Performer and Event Planner Details on Profile Page
Added the display of performer and event planner display on the profile page of a user. All of the user's performer data is displayed for each type of performer that they are. Planning on creating a page where more specific data such as past events and upcoming events, both for EventPlanner users and Performer users