Made some changes and added some things on existing pages; added "outlines" of new pages; added navigation bar and some other components
Added display of the names of past and upcoming events on the Profile Page in the EventPlannerProfileOverview component.

Added a Navigation Bar component that allows the user to switch between their Profile Page, Requests Page, Messages Page and Connections Page - the last three listed have yet to be implemented.

Added a SlotDetailsForm component that allows an event planner to add slots to their event for performers to be given to perform in.

Made it so that events can be created and added to the database on the CreeateEventPage.

Added the ability to send requests to other Event Planner users to co-host an event.

Changed the registration system. The details are no longer added to the database after finishing filling in and submitting the information on the Registration Page. That data that was submitted is now passed to the Details Page and all the data is added at once to the database.

Changed the system of navigating to the Profile Page. The email address of the user whose profile page is being navigated to is passed to the page from the page being navigated from. If the email being passed equals the email in session storage (i.e. the user's email), then the profile page displayed is the user's, otherwise it is another user's profile page.

Added buttons for creating events and for creating groups that take you to the respective pages to do so. These buttons only display if the user is on their own profile page.

Added a button for sending a friend request to a user that displays when the user is on someone else's profile page and added functionality to this button.

Changed the background colour of the vertical panel on the Profile Page.

Added the moving of events from upcoming events to past events if the end date of the upcoming event is after the current time. This process is triggered whenever someone's profile is loaded.

Created files for the Create Group, Connections, Messages, and Requests pages. These pages have not yet been implemented, however, but they can still be navigated to.