Connections Page, Global Styles, Small Fixes and Edits
Worked more on the Connections page - still have to add the sending of requests. To do this, a PageContainer CSS component was added to each page and each page was wrapped in this.

Added a background image to the website and made the actual pages display smaller.

Fixed an error on the CreateEventPage where it was not compatible with the new database schema.