Continued work on Connections Page
Fixed some issues and bugs I was dealing with. You can now select an event to choose a performer for, and if that event has slots, you can choose a slot to find a performer for. The part of finding the performers has yet to be implemented, but that won't take too long - I already am mostly done with creating the PerformerConnectionsDisplay component.