Reviews, Friend Displays, Connections Page Bug
Finished implementing the feature of being able to add reviews and view them on a user's profile page.

Implemented the display of friends of a user of their profile page.

Fixed a bug on the Connections Page where nothing was displaying if a user was only an event planner or only a performer.